requirements.txt
flask~=2.0.1
Werkzeug~=2.0.1
Flask-Session
mysql-connector-python
